refactor: extract community constants and utilities
